// src/app/modules/auth/models/UserModel.ts
import { AuthModel } from './AuthModel'
import { UserAddressModel } from './UserAddressModel'
import { UserCommunicationModel } from './UserCommunicationModel'
import { UserEmailSettingsModel } from './UserEmailSettingsModel'
import { UserSocialNetworksModel } from './UserSocialNetworksModel'

export interface UserModel {
    id: number
    username: string
    password: string | undefined
    email: string
    cryp_id: number | undefined
    strategy: number | undefined
    firstName: string
    referral: string
    lastName: string
    type: string
    childs: Array<number>
    parent: number | undefined
    companyName?: string
    phone?: string
    roles?: Array<number>
    pic?: string
    language?: 'en' | 'de' | 'es' | 'fr' | 'ja' | 'zh' | 'ru'
    timeZone?: string
    website?: 'https://keenthemes.com'
    emailSettings?: UserEmailSettingsModel
    auth?: AuthModel
    communication?: UserCommunicationModel
    address?: UserAddressModel
    socialNetworks?: UserSocialNetworksModel
}